Package update and allow setting prefered installer
Fixed #144
