+++ src/Winget.CommunityRepository.Ef/Winget.CommunityRepository.Ef.csproj
@@ -20,7 +20,9 @@
     <NoWarn>1701;1702;CS8618</NoWarn>
   </PropertyGroup>
 	<ItemGroup>
-		<PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="8.0.1" />
+		<PackageReference Include="Microsoft.Extensions.Caching.Memory" Version="8.0.1" />
+		<PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="8.0.2" />
+		<PackageReference Include="System.Text.Json" Version="8.0.5" />
 	</ItemGroup>
   <ItemGroup Condition=" '$(TargetFramework)' == 'net8.0' ">
     <PackageReference Include="Microsoft.EntityFrameworkCore.Design" Version="8.0.4">
