+++ tests/WingetIntune.Tests/Intune/IntuneManagerTests.cs
@@ -1,6 +1,7 @@
 ﻿using Microsoft.Extensions.Logging.Abstractions;
 using System.Runtime.InteropServices;
 using Winget.CommunityRepository.Models;
+using WingetIntune.Commands;
 using WingetIntune.Interfaces;
 using WingetIntune.Models;
 
@@ -11,7 +12,7 @@ public class IntuneManagerTests
     [Fact]
     public async Task GenerateMsiPackage_OtherPackage_ThrowsError()
     {
-        var intuneManager = new IntuneManager(null, null, null, null, null, null, null, null, null, null);
+        var intuneManager = new IntuneManager(null, null, null, null, null, null, null, null, null, null, new ComputeBestInstallerForPackageCommand());
         var tempFolder = Path.Combine(Path.GetTempPath(), "intunewin");
         var outputFolder = Path.Combine(Path.GetTempPath(), "packages");
 
@@ -80,7 +81,7 @@ The Azure command-line interface (Azure CLI) is a set of commands used to create
 
         var intunePackager = Substitute.For<IIntunePackager>();
 
-        var intuneManager = new IntuneManager(new NullLoggerFactory(), fileManager, processManager, null, null, null, intunePackager, null, null, null);
+        var intuneManager = new IntuneManager(new NullLoggerFactory(), fileManager, processManager, null, null, null, intunePackager, null, null, null, new ComputeBestInstallerForPackageCommand());
 
         await intuneManager.GenerateInstallerPackage(tempFolder, outputFolder, IntuneTestConstants.azureCliPackageInfo, new PackageOptions { Architecture = Models.Architecture.X64, InstallerContext = InstallerContext.User }, CancellationToken.None);
 
@@ -105,7 +106,7 @@ The Azure command-line interface (Azure CLI) is a set of commands used to create
         fileManager.DownloadFileAsync($"https://api.winstall.app/icons/{packageId}.png", logoPath, null, false, false, Arg.Any<CancellationToken>())
             .Returns(Task.CompletedTask);
 
-        var intuneManager = new IntuneManager(new NullLoggerFactory(), fileManager, null, null, null, null, null, null, null, null);
+        var intuneManager = new IntuneManager(new NullLoggerFactory(), fileManager, null, null, null, null, null, null, null, null, new ComputeBestInstallerForPackageCommand());
         await intuneManager.DownloadLogoAsync(folder, packageId, CancellationToken.None);
 
         //call.Received(1);
@@ -137,7 +138,7 @@ The Azure command-line interface (Azure CLI) is a set of commands used to create
         fileManager.DownloadFileAsync(packageInfo.InstallerUrl.ToString(), installerPath, hash, true, false, Arg.Any<CancellationToken>())
             .Returns(Task.CompletedTask);
 
-        var intuneManager = new IntuneManager(new NullLoggerFactory(), fileManager, null, null, null, null, null, null, null, null);
+        var intuneManager = new IntuneManager(new NullLoggerFactory(), fileManager, null, null, null, null, null, null, null, null, new ComputeBestInstallerForPackageCommand());
         await intuneManager.DownloadInstallerAsync(folder, packageInfo, CancellationToken.None);
 
         //call.Received(1);
